[Chore]: 바뀐 공홈 크롤링 API 요구사항에 맞춰 변경
