+++ src/main/java/org/sopt/makers/internal/dto/sopticle/SopticleVo.java
@@ -1,14 +1,6 @@
 package org.sopt.makers.internal.dto.sopticle;
 
 public record SopticleVo(
-        String link,
-        SopticleUserVo author
+        String link
 ) {
-    public record SopticleUserVo(
-            Long id,
-            String name,
-            String profileImage,
-            Integer generation,
-            String part
-    ) {}
 }
